configure.ac: check for Qt5 version first
Despite zbar can be compiled with either Qt4 or Qt5, the
current logic seeks only Qt4. Change it to try first Qt5.

Yet, as people may have both but want to build it with
Qt4, add a --without-qt5 configuration option that will
make it to check only for Qt4, just like before.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>