zbarcam: add qt and gtk variants
The original zbarcam use standard X11 libraries to display
video. That provides an ugly interface, and could be
slow on high resolutions.

As Zbar has already test examples for qt and gtk, let's
promote them to real applications. One advantage is that
both qt and gtk libraries will be tested when building
those.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>