v4l2: add fallback for systems without v4l2_ext_controls which field
The v4l2_ext_controls.which field was introduced on Kernel 4.4,
in order to solve some ambiguities and make easier to handle
controls.

Yet, there are several systems running older Kernels. As the
newer Linux Kernels are backward-compatible with the old way,
we can change the logic in a way that would allow someone to
build it against a kernel > 4.4, while letting it to keep running
with legacy Kernels.

Reported-by: James Hilliard <james.hilliard1@gmail.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>