configure.ac: The pdf417 code is incomplete. Warn about that
While it seems that all the image processing part for pdf417
is there, the code misses the part with it would be infering
dimensions and producing the output code, as warned inside
the code:

    /* FIXME TBD infer dimensions, save codewords */

Due to that, enabling it won't actually work, and should be
used only by some developer that are willing to complete
its implementation. Document that.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>