barcodetest.py: make it use the new ean-13.png barcode
The ean-13 file is hardcoded on it. Change it to use the
new one found at example, and point to a fallback address
via https at linuxtv.org.

It should be noticed that https is only supported with
ImageMagick. So, add a warning if the build is using
GraphicsMagick instead.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>