Silently ignore --nodbus argument
If ZBar is built without dbus support, make test and any other
user script that might be using this option would fail.

So, let's just silently ignore it, in order to make the
scripts more portable.

While here, move the dbus option output to the right place,
instead of outputing it as if it were a warning.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>