test_decode: better handle decoding errors
Sometimes, the random generated images may be decoded by
multiple decoders. Handle those as warnings, and keep
waiting for the right code, only reporting an error when
either a symbol was missing with the right code or when
it was badly decoded.

With that, "make regress" now produce a more coherent
result:

    [28417] SEED=-1466011001: warning: expecting DataBar, got EAN-2
    [35442] SEED=-591992353: ERROR: MISSING CODE-39 (YP%CC-.)
    [64558] SEED=1906143625: ERROR: MISSING CODE-128 (41714534408557)
    [79042] SEED=-1776449378: ERROR: MISSING CODE-39 (Z5EKHI5I6)
    [84522] SEED=-966054662: warning: expecting DataBar, got EAN-2
    [98350] SEED=1992973531: warning: expecting DataBar, got EAN-2
    decoder FAILED (3 errors).

Calling it with one of those seeds will also present a better
result:

	test_decode -s -1466011001
	[0] SEED=-1466011001
	testing data: 58022598404365
	[0] SEED=-1466011001: warning: expecting DataBar, got EAN-2
	testing alpha: kqw9:'{,sx
	decoder PASSED.

On this specific case, it decoded as data=10 (EAN-2) and with
data=58022598404365(DataBar) as expected.

The first one was an expurious decoding.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>