test_decode: use argp to decode parameters
The existing argument parsing logic is weird: if parameters
are passed out of an expected order, it may be ignored. Also,
it doesn't provide any help about what parameters are available.

Use argp instead, with simplifies the code and provide a
help about the expected parameters.

As a bonus, the code inside the main() function is now
clearer and easier to understand. It also allows to combine
-n with -s, with could be useful to start a regression test
from a different seed.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>