test_decode: add a tolerance margin for errors
As the testset generated by the regression tool may contain
noise, it is possible that some images will cause some
confusion, causing them to not be decoded, or causing
another decoder to think it is a valid image for them.

So, tolerate spurious detections and missing detection, if
is is less than 0.01% of the sample set.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>