ean: ISBN-13 has precedence
Physically, ISBN-13 and ISDN-10 codes are identical, and
are just EAN-13 with special meaning.

ISBN-10 starts with "978"; ISDN-13 may start with either
"978" or "979". The check digit label is different for
ISDN-10.

Anyway, as, ISDN-13 is the standard, since 2007, it makes
sense to make it the default.

It should be noticed that, by default, ZBar won't enable
ISDN-13. So, if one needs to decode EAN-13 as ISDN, he
needs to explicitly enable it:

	zbarcam --set isbn10.enable

As, otherwise, the code will be decoded as EAN-13.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>