move d-bus code to img_scanner.c
While parsing D-Bus messages at processor.c works fine for
zbarcam and zbarimg apps, this is not generic enough. Making
it to work, for example, with zbarcam-qt, would be really
painful.

So, while not too late, move the code to be inside img_scanner,
where it belongs.

While here, don't make the API/ABI dependent if HAVE_DBUS or
not, as this may cause problems in the future.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>