configure.ac: fix an error at libv4l2 package check
Writing m4 logic is hard, as it is required to properly
mark begin/end of each block.

In the case of the libv4l detection check, there's a missing
[] for PKG_CHECK_MODULES(). This causes it to produce a wrong
code, ending by having 2 errors when the script runs:
	checking for libv4l2.h... yes
	./configure: line 15166: V4L2_CFLAGS: command not found
	./configure: line 15167: V4L2_LIBS: command not found

While here, better ident the blocks and fix the error message
when libv4l2.so is not found.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>