symbol.c: fix symbol hash logic
ZBar uses an array (scanner sym_configs) in order to store
some decoder-specific configurations (currently,
only ZBAR_CFG_UNCERTAINTY is stored there).

As the value of zbar_symbol_type_e can be up to 0x700,
it uses a poor man's hash table to map it into an array with
NUM_SYMS elements (19 elements).

The idea of using a hash to convert from zbar_symbol_type_e
into an array index between 0 and NUM_SYMS - 1 is interesting,
but it has a drawback: the latest added symbology (SQCODE)
caused a collision, making one of the used vars to be a -1.

As the hash code there doesn't provide any collision detection
mechanism, it checks for problems with assert() calls.

So, if one tries to set or retrieve ZBAR_CFG_UNCERTAINTY config
for sqcode, it gets an abort().

While we could come up with some improved hash logic there that
would avoid collisions, as the range of symbol codes that makes
sense for that code is not bigger than 128, let's just use an
static array with 129 elements (instead of the original one
with 32 elements).

I opted to preserve the values returned by the past hash
table code for the symbologies that makes sense.

In the case of the new SQCODE, I opted to set it to 1
(it could also be 0).

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>