scan_video: improve logic to remove duplicated/invalid devnodes
The same devnode can appear with different names. Also,
modern Kernels have an extra video node for metadata.

So, better to validate if we're opening the right one,
and cleanup duplicated ones.

As I wrote a code for that for camorama, import my code,
adapting it to work with zbar.

While the main focus here is to improve zbarcam-qt,
I opted to keep the code in C, as at the original code,
as it would allow its usage on other programs.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>