Add an API to allow getting supported resolutions and change them
Export the detected resolutions to ZBar API.

Changing resolutions in runtime can be tricky with V4L2 and
ZBar, as it needs to dequeue all pending buffers, in order
to free them and reallocate.

While V4L2 allows to do that without needing to close/reopen
the file descriptor, with current ZBar implementation, it is
a lot easier if we just close/reopen, just like if we were
opening a different device.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>