Fix Java detection and fix it for newer Java versions
There's a mess at configure.ac at the java detection code,
and on its usage at check-java target.

Place the Java detection code in just one place, reorganizing
the code in a way that it will use the canonical name of the
javac directory as default, in the case JAVA_HOME is not
explicitly set. It will also use the detected JAVA_HOME at

	make check-java

target.

Also, java versions bigger than 8 don't come anymore with
javah. Instead, javac was incremented to produce headers,
but  with a different syntax and more limited features, as
it doesn't allow anymore to generate a single header file
for multiple java sources.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>