configure.ac: avoid calling PKG_CHECK_MODULES twice for Magick
According with its documentation:
	https://autotools.io/pkgconfig/pkg_check_modules.html

we can't call PKG_CHECK_MODULES twice with the same prefix.
The only exception is when it is called inside a chained
call.

Well, we do that on Magick, when GraphicsMagick is selected.

Change the logic to avoid that, as this can cause random
hard to track errors.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>