Makefile.am: split build on an independent Makefile.am for zbar and gtk
As we're about to add support for GIR build under gtk/dir with
requires it to have its own Makefile.am, use SUBDIRS instead of
include for it.

Changing it would cause linkage breakages with Ubuntu, as we need
to install first libzbar, then libgbargtk, and finally any other
stuff that depends on them.

So, we also need to split the zbar/Makefile.am.inc into its own
makefile.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>