zbargtk: get rid of gdk_threads for good
The usage of gdk_threads() to handle image processing thread
makes it a lot harder to have a version that would be compatible
with both Gtk3 and Gtk2. Also, it would require changes at
userspace.

So, instead, use g_idle() to handle the state change, with is
notified via asynchronous messages, plus the image buffer
handling.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>