debian: shut up dpkg-gensymbols
Travis CI is crashing due to a new symbol it doesn't recognize.

Add the new versions to changelog with a dummy message and
add the new symbol at debian/libzbargtk0.symbols just to
shut up TravisCI.

The real debian log should actually be at the Debian packaging
system and not here.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>