zbargtk.h: make it capable of producing a good ZBar .gir file
Making it properly parseable by g-ir-scanner is not easy:
there are several tricks. Add the proper annotations for it to
produce a working .typelib and .gir file.

Note: the main trick that the parser should be able to recognize
zbar_gtk_get_type(), in order to generate a class. It is not
possible to make it produce a class called ZBarGtk, as I would
like, so, the next best good thing is to recognize it as ZBar.Gtk.

The reason is that the GIR logic will strip part of the name of
the function, but it needs a final name for the class. So, if it
strips "zbar_gtk_", the name would be "get_type", with is invalid.

By stripping only "zbar_" prefix, the name will be "gtk_get_type",
and the class will be called "Gtk", inside the "ZBar" namespace.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>