configure.ac: allow specify python version via --with-python
We have already a parameter to enable/disable python. As we
now support python 2/3 autodetection, change the logic here
to use what's defined there in order to decide what python
version will be used at the build.

For now, keeps the default to python2. Future versions will
change it to "auto".

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>