zbargtk: Fix a Gtk3 port bug causing it to crash
I was stuck getting this error with Gtk3:
	Gtk:ERROR:gtkwidget.c:5531:gtk_widget_unrealize: assertion failed: (!widget->priv->mapped)

The documentation for realize/unrealize is very poor, so I
had to seek for other places where unrealize were implemented.

After several try and errors, I finally found a place with
a simple but effective solution:

All I had to do were to unset the GTK_MAPPED flag, and
the bug was gone!

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>