gtk: get rid of gtk-gir-dump.c
g-ir-scanner can work with either a library or a program.

During the development, it only started working after using
the program mode. Now that GIR is working, it is *also*
working directly from the library.

Using the library instead of a program have several
advantages. The main one is that the setup becomes a lot
clearer. It should also be simpler to make it working on
more complex build scenarios.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>