gtk/Makefile.am: fix a bug with out-of-tree Zbar-1.0.gir generation
At least with Fedora 30 builds using mock (e. g. inside a chroot and
having the build dir different than the source dir, using
gobject-introspection-1.60, there is a bug with GIR file generation:
sometimes, g-ir-scanner is not capable of producing a C file that would
be loading libzbargtk.so.0. So, it fails with:

    error while loading shared libraries: libzbargtk.so.0: cannot open shared object file: No such file or directory

I suspect that it has something to do with libtool-2.4.6.

The fix is hackish, but it should be safe: it should manually include
the paths where libtool generate those at INTROSPECTION_SCANNER_ARGS.

It should be noticed that, this shouldn't affect a non-buggy environment,
so, better to be safe than sorry.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>