test/test_gi.py: get rid of Gtk3 warnings
Change the code to avoig Gtk3 warnings. We don't care too
much about Gtk2 here, as Gtk2 as issues with GI. Yet, I
suspect that none of those changes will break with the latest
Gtk2 versions.

Among the changes is the import of GdkPixbuf. Without this,
opening an image won't work.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>