travis: prepare it to add windows build with MinGw-w64
As ZBar can also be built on Windows with MinGw-w64,
prepare the build logic to work with Windows.

The Windows preparation scripts use  msys2 in order to do
install autoconf/automake and do the ZBar's windows build.

For now, keep the build disabled, as there are still some
issues that will require other changes for it to work.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>