makefiles: fix CPPFLAGS
a recent change in order to support native Windows build caused
ZBar to break on some builds due to the lack of the proper
CPPFLAGS. While debugging it, I noticed that the definition at
configure.ac assumes a single Makefile.am file. Because of that,
it was adding a "./include" on some targets at gtk and zbar
directories.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>