README-windows.md: update to reflect the current status
Right now, we're not providing a built version for windows
anywhere. Also, the library information there are bound to
the versions that used to be available at the time ZBar was
maintained at SourceForge.

Make the library text more flexible by removing the specific
versions, update the site informations and making clearer
about the license terms that apply.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>