config: add boolean option for raw binary decoding
When decoders encounter binary data, they attempt to guess the encoding
of the data and convert it to text. This destroys other types of data.

The new boolean decoder configuration option ZBAR_CFG_BINARY
suppresses this automatic character set conversion.
The decoders that support it will output the raw binary data.

Signed-off-by: Matheus Afonso Martins Moreira <matheus.a.m.moreira@gmail.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>