zbarcam: add one shot scanning mode
Continuous scanning can cause problems if users expect a single barcode.
There could be multiple barcodes in an image.
The same barcode could be detected and decoded twice.
The result is multiple outputs or duplicated data in the output.
Programs that consume this output must account for this.

With the new one shot mode, zbarcam will exit successfully
after scanning exactly one barcode. This simplifies use of the tool.

Suggested-by: Dominik Xaver Hörl <hoe.dom@gmx.de>
Signed-off-by: Matheus Afonso Martins Moreira <matheus.a.m.moreira@gmail.com>