travis: debian-sid: create a tarball for pbuilder
It seems that, on previous versions (either at TravisCI
or at sid), the tarball with the sources were either
already present at the parent dir or were created by
pbuilder.

Well, this is not the case anymore, so let's manually
create it at the travis script.

This should restore the debian sid build at TravisCI.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>