travis: debian-sid: update debian directory
The contents of the debian sid build directory doesn't reflect
upstream anymore: it still builds against python2, with is now
deprecated.

So, let's re-sync it from today's upstream debian file, as
found at:

	http://deb.debian.org/debian/pool/main/z/zbar/zbar_0.23-1.3.debian.tar.xz

It should be noticed that this file won't be building Java, and
may also be missing some features, as packages are now different
than what it used to be.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>