travis: change windows building logic
The windows build is very hard to get it right, as TravisCI
has an incompatible toolset that needs to be first removed
and then MSYS2 has to be manually installed.

It doesn't help that autotools try to search for other
tools under /usr/bin directory.

So, there are some magic here.

Yet, this is not enough for making it run clearly,
but at least it is following the the recommended
way to work with MSYS2, as stated at:

	https://docs.travis-ci.com/user/reference/windows/

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>