travis: fix native Windows build
With msys2, there's no need to use and target on configure.
Also, the ones used there are now wrong.

The prefix is also wrong: it causes autoconf to crash.

As we don't really matter where it will be installed, let's
just use, instead, the msys2 prefix.

We also need to re-enable the CONFIG_SHELL, as this is
called by make. By default, it would try to use a shell
under "Program Files", with doesn't work, due to the
white spaces.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>