travis: don't allow failures on Windows anymore
As the Windows build now return success, we don't need to
set them as "allow failures".

It should be noticed, however, that "make install" still
fails on Windows, because the install procedure is not
capable of creating some directories.

Perhaps we need to add "sudo" or something like that to it,
or install under "Program Files", but that would require some
extra logic because of the whitespaces. So, for now, let's
just ignore errors on make install, as we're more interested
on checking that the building phase is OK.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>