zbarimg: better handle output mode parameters
The --xml, --noxml and --raw parameters affect how the output
is done.

However, right now, they're positional, with causes some weird
behaviour, for example, if "--xml" is used at the end of the
line.

This doesn't make much sense. So, let's change the logic, by
setting either one of the three modes for all images.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>