workflows: on_release: use debuild instead of pbuilder
While pbuilder helps to solve some things and to create a
clean build environment, the current procedure has some
issues with regards to version numbering: basically, it
expects that the ZBar version to be the one from the Ubuntu
repository (0.23.1).

While this is ok for CI bots, it will cause issues if minor
changes. The package names will also be weird.

So, let's use debuild instead. This requires a small trick,
though:

dh_shlibdeps will start complaing for each new symbol, due to
SONAME debian rules. So, we need to patch debian rules to
make it ignore them.

Also, the script needs to ensure that the same version number
as defined at configure.ac will be at debian/changelog.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>