qrdectxt: use a define for the enc list size
The enc_list has 4 elements, but it calls a function that
defines the expected range as having just 3 elements.

The actual code is correct, as internally it uses 4 as its
size, but that's confusing, as reported by:

https://github.com/mchehab/zbar/issues/114

So, define a macro with the size, and use it on all for()
loops and all array declarations.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>