Bump version to 0.23.92
This version is a pre-release version. The only difference is
that it now brings a --enable-static, meant to help distributing
Windows binaries.

Unfortunately, the only way to test it is by creating a new
release, as we depends on the github workflows to create the
binary, which is trigged only for new releases.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>