Updated the Python GTK widget to Python 3 and GTK 3.
The previous implementation of the Python wrapper around the GTK widget
was written for Python 2 and GTK 2.
Python 2 has been EOL for more than two years, but the widget is still
very interesting.

Also, the way to integrate GTK and Python has changed a lot during the
years. So, the ZBarGtk widget type is registered automatically by gi,
and the only needed step in the bindings is creating and returning
the widget somehow.

I also implemented the request video size functionality through new
attributes: video-width and video-height.

The only feature the Python wrapper is missing is the possibility to
scan a GdkPixbuf, because I could not find a way to register a new
method to the __gi__.ZBarGtk type without reimplementing it completely.

Signed-off-by: Pier Angelo Vendrame <vogliadifarniente@gmail.com>